Check if KStream for an input topic already exists and re-use it.
